// src/api-client/models/employee-response-dto.ts
/* tslint:disable */

/**
 * API Documentation
 * API - IT Asset Tracker
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface EmployeeResponseDto
 */
export interface EmployeeResponseDto {
  /**
   * Unique employee identifier
   * @type {number}
   * @memberof EmployeeResponseDto
   */
  id: number;
  /**
   * Employee full name
   * @type {string}
   * @memberof EmployeeResponseDto
   */
  fullName: string;
  /**
   * Unique employee code
   * @type {string}
   * @memberof EmployeeResponseDto
   */
  employeeCode: string;
  /**
   * Employee email address
   * @type {string}
   * @memberof EmployeeResponseDto
   */
  email: string;
  /**
   * When the record was created
   * @type {string}
   * @memberof EmployeeResponseDto
   */
  createdAt: string;
  /**
   * Soft-delete timestamp; null while the record is active
   * @type {string}
   * @memberof EmployeeResponseDto
   */
  deletedAt: string | null;
}